requirements.txt
numpy
datasets
transformers
sentence-transformers
tldextract
deepspeed
accelerate
polars
spacy
wandb
fasttext
rapidfuzz
clean-text[gpl]
deep-translator
flair
faiss-gpu
gensim
nltk
rank-bm25